# src/swagger/swagger.yaml
basePath: /
definitions:
...
  api.InfoCreateRequest:
    properties:
      broadcasts:
        allOf:
        - $ref: '#/definitions/whatsapp.WhatsappBoolean'
        description: should handle broadcast messages
      calls:
        allOf:
        - $ref: '#/definitions/whatsapp.WhatsappBoolean'
        description: should handle calls
      devel:
        description: enable debug mode (devel)
        type: boolean
      groups:
        allOf:
        - $ref: '#/definitions/whatsapp.WhatsappBoolean'
        description: should handle groups messages
      readreceipts:
        allOf:
        - $ref: '#/definitions/whatsapp.WhatsappBoolean'
        description: should emit read receipts
      readupdate:
        allOf:
        - $ref: '#/definitions/whatsapp.WhatsappBoolean'
        description: should send markread requests when receiving messages
    type: object
  api.InformationResponse:
    properties:
      debug:
        description: Extra interface{} `json:"extra,omitempty"`
        items:
          type: string
        type: array
      server:
        $ref: '#/definitions/models.QpWhatsappServer'
      status:
        type: string
      success:
        type: boolean
      version:
        type: string
    type: object
...
  models.QpDispatching:
    properties:
      broadcasts:
        allOf:
        - $ref: '#/definitions/whatsapp.WhatsappBoolean'
        description: should handle broadcast messages
      calls:
        allOf:
        - $ref: '#/definitions/whatsapp.WhatsappBoolean'
        description: should handle calls
      connection_string:
        description: destination URL (webhook) or connection string (rabbitmq)
        type: string
      extra:
        description: extra info to append on payload
      failure:
        description: first failure timestamp
        type: string
      forwardinternal:
        description: forward internal msg from api
        type: boolean
      groups:
        allOf:
        - $ref: '#/definitions/whatsapp.WhatsappBoolean'
        description: should handle groups messages
      readreceipts:
        allOf:
        - $ref: '#/definitions/whatsapp.WhatsappBoolean'
        description: should emit read receipts
      readupdate:
        allOf:
        - $ref: '#/definitions/whatsapp.WhatsappBoolean'
        description: should send markread requests when receiving messages
      success:
        description: last success timestamp
        type: string
      timestamp:
        type: string
      trackid:
        description: identifier of remote system to avoid loop
        type: string
      type:
        description: webhook or rabbitmq
        type: string
    type: object
...
  models.QpResponse:
    properties:
      debug:
        description: Extra interface{} `json:"extra,omitempty"`
        items:
          type: string
        type: array
      status:
        type: string
      success:
        type: boolean
    type: object
...
  models.QpTimestamps:
    properties:
      event:
        description: Last received event timestamp
        type: string
      message:
        description: Last received message timestamp
        type: string
      start:
        description: Server start timestamp
        type: string
      update:
        description: Last database update timestamp
        type: string
    type: object
...
  models.QpWhatsappServer:
    properties:
      broadcasts:
        allOf:
        - $ref: '#/definitions/whatsapp.WhatsappBoolean'
        description: should handle broadcast messages
      calls:
        allOf:
        - $ref: '#/definitions/whatsapp.WhatsappBoolean'
        description: should handle calls
      devel:
        type: boolean
      dispatching:
        items:
          $ref: '#/definitions/models.QpDispatching'
        type: array
      groups:
        allOf:
        - $ref: '#/definitions/whatsapp.WhatsappBoolean'
        description: should handle groups messages
      readreceipts:
        allOf:
        - $ref: '#/definitions/whatsapp.WhatsappBoolean'
        description: should emit read receipts
      readupdate:
        allOf:
        - $ref: '#/definitions/whatsapp.WhatsappBoolean'
        description: should send markread requests when receiving messages
      reconnect:
        description: should auto reconnect, false for qrcode scanner
        type: boolean
      timestamp:
        type: string
      timestamps:
        $ref: '#/definitions/models.QpTimestamps'
      token:
        description: Public token
        maxLength: 100
        type: string
      user:
        maxLength: 36
        type: string
      verified:
        type: boolean
      wid:
        description: Whatsapp session id
        maxLength: 255
        type: string
    type: object
...
  whatsapp.WhatsappBoolean:
    enum:
    - -1
    - 0
    - 1
    type: integer
    x-enum-varnames:
    - FalseBooleanType
    - UnSetBooleanType
    - TrueBooleanType
...
info:
  contact:
    name: QuePasa Support
    url: https://github.com/nocodeleaks/quepasa
  description: QuePasa is a Go-based WhatsApp bot platform that exposes HTTP APIs
    for WhatsApp messaging integration
  license:
    name: GNU Affero General Public License v3.0
    url: https://github.com/nocodeleaks/quepasa/blob/main/LICENSE.md
  termsOfService: https://github.com/nocodeleaks/quepasa
  title: QuePasa WhatsApp API
  version: 4.0.0
paths:
...
  /info:
    delete:
      consumes:
      - application/json
      description: Delete bot/server information and settings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.InformationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete bot information
      tags:
      - Information
    get:
      consumes:
      - application/json
      description: Get bot/server information and settings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.InformationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Get bot information
      tags:
      - Information
    patch:
      consumes:
      - application/json
      description: Update bot/server information and settings
      parameters:
      - description: Settings update
        in: body
        name: request
        schema:
          properties:
            settings:
              type: object
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.InformationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Update bot information
      tags:
      - Information
    post:
      consumes:
      - application/json
      description: Create a new bot/server with configuration before QR code scanning
      parameters:
      - description: Server creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.InfoCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Server updated
          schema:
            $ref: '#/definitions/api.InformationResponse'
        "201":
          description: Server created
          schema:
            $ref: '#/definitions/api.InformationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.QpResponse'
      security:
      - ApiKeyAuth: []
      summary: Create bot configuration
      tags:
      - Information
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-QUEPASA-TOKEN
    type: apiKey
swagger: "2.0"